Initial version capable of connecting to the Mumble server and streaming audio from a local MPD's FIFO output.
It cannot even gracefully quit yet, and cannot comminicate to MPD,
detect users present in channel or bring you a bottle of beer.
It just tries to join the server and starts streaming audio
if the connection was successful.

Example config is not included because I don't yet know how it will look like.
See the source code for configuration parameters names.